+++ remember_frontend/ModuleTest.jsx
@@ -8,7 +8,7 @@ import { StyleSheet, Text, View, ScrollView, Image } from 'react-native'
 import { SafeAreaView, SafeAreaProvider } from "react-native-safe-area-context"
 import * as Device from "expo-device"
 import Constants from "expo-constants"
-import ImagePicker from "./modules/ImagePicker"
+import ImageUpload from "./modules/ImageUpload"
 
 
 
@@ -41,7 +41,7 @@ const ModuleTest = () => {
         }}
         alt="it doesnt load"/>
       <Video />
-      <ImagePicker />
+      <ImageUpload />
     </ScrollView>
     </SafeAreaView>
     </SafeAreaProvider>
