+++ remember_frontend/package.json
@@ -18,6 +18,7 @@
     "expo-constants": "^17.0.3",
     "expo-device": "^7.0.1",
     "expo-image": "^2.0.3",
+    "expo-image-picker": "^16.0.3",
     "expo-location": "^18.0.4",
     "expo-module-scripts": "^4.0.2",
     "expo-notifications": "~0.29.11",
